FIX: parameter type 잘못 명세된 부분 수정

+++ src/main/java/com/adinstar/pangyo/mapper/PostMapper.java
@@ -13,5 +13,5 @@ public interface PostMapper {
     Post selectById(long id);
     int insert(Post post);
     int update(Post post);
-    int delete(Integer postId);
+    int delete(long postId);
 }
